[FLINK-28053][sql-gateway] Introduce operation lock to execute operation in sequence
